requirements.txt
pygame-ce
pygame_gui
